Wasan's contribution reflection.md
hello, i added my own reflection + i edited the spacing a little. i think we have to double space if we want our paragraph to be on a new line. (e.g if i wanted "Wasan's reflection" on one line because its the header and my paragraph to be underneath it). I'm going to do a pull request just so i can see if my theory is correct, and if I am, I will merge to main branch.